requirements.txt
websockets>=12.0
python-dotenv>=1.0.0
aiohttp>=3.11.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
logging>=0.4.9.6
typing-extensions>=4.9.0
numpy>=1.24.0
black>=24.0.0
pylint>=3.0.0
autopep8>=2.0.0